+++ lib/Auth/Process/IsCesnetEligible.php
@@ -9,26 +9,34 @@
  */
 class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_ProcessingFilter
 {
-	const UNIVERSITY = "university";
-	const AVCR = "avcr";
-	const LIBRARY = "library";
-	const HOSPITAL = "hospital";
-	const OTHER = "other";
-	const EDUID_IDP_GROUP = "http://eduid.cz/uri/idp-group/";
+
+	const CONFIG_FILE_NAME = 'module_cesnet_IsCesnetEligible.php';
+	const ORGANIZATION_LDAP_BASE = 'ou=Organizations,o=eduID.cz,o=apps,dc=cesnet,dc=cz';
 
 	const HOSTEL_ENTITY_ID = "https://idp.hostel.eduid.cz/idp/shibboleth";
 
 	const INTERFACE_PROPNAME = "interface";
 	const CESNET_ELIGIBLE_LAST_SEEN_ATTR = "cesnetEligibleLastSeenAttr";
 	const DEFAULT_ATTR_NAME = 'isCesnetEligibleLastSeen';
+	const LIST_OF_PERUN_ENTITY_IDS = 'listOfPerunEntityIds';
 
 	private $cesnetEligibleLastSeen;
 	private $cesnetEligibleLastSeenAttr;
 
-	private $metadata;
-	private $entityCategory;
+	private $spEntityId;
+	private $idpEntityId;
 	private $eduPersonScopedAffiliation = array();
+	private $listOfPerunEntityIds = array();
 
+	/**
+	 * @var sspmod_perun_LdapConnector
+	 */
+	private $cesnetLdapConnector;
+
+	/**
+	 * @var sspmod_perun_RpcConnector
+	 */
+	private $rpcConnector;
 
 	public function __construct($config, $reserved)
 	{
@@ -44,36 +52,28 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
 			$this->attrName = self::DEFAULT_ATTR_NAME;
 		}
 
+		if (isset($config['listOfPerunEntityIds'])) {
+			$this->listOfPerunEntityIds = $config['listOfPerunEntityIds'];
+		}
+
 		$this->cesnetEligibleLastSeenAttr = $config[self::CESNET_ELIGIBLE_LAST_SEEN_ATTR];
+
+		$this->cesnetLdapConnector = (new sspmod_perun_AdapterLdap(self::CONFIG_FILE_NAME))->getConnector();
+		$this->rpcConnector = (new sspmod_perun_AdapterRpc())->getConnector();
+
 	}
 
 	public function process(&$request)
 	{
 		assert('is_array($request)');
 		$user = $request['perun']['user'];
-
-		$this->metadata = SimpleSAML_Metadata_MetaDataStorageHandler::getMetadataHandler();
-		$sourceIdpMeta = $this->metadata->getMetaData( $request['saml:sp:IdP'], 'saml20-idp-remote');
-
-		if (isset($sourceIdpMeta['EntityAttributes']['http://macedir.org/entity-category'])) {
-			$entityCategoryAttributes = $sourceIdpMeta['EntityAttributes']['http://macedir.org/entity-category'];
-		} else {
-			SimpleSAML\Logger::error("cesnet:IsCesnetEligible - There are no element with name 'EntityAttributes' "
-				. "and subelement with name 'http://macedir.org/entity-category' in metadata for IdP with entityId "
-				. $request['saml:sp:IdP'] . "!");
-			$entityCategoryAttributes = array();
-		}
+		$this->spEntityId = $request['SPMetadata']['entityid'];
+		$this->idpEntityId = $request['saml:sp:IdP'];
 
 		if (isset($request['Attributes']['eduPersonScopedAffiliation'])) {
 			$this->eduPersonScopedAffiliation = $request['Attributes']['eduPersonScopedAffiliation'];
 		} else {
 			SimpleSAML\Logger::error("cesnet:IsCesnetEligible - Attribute with name 'eduPersonScopedAffiliation' did not received from IdP!");
-		}
-
-		foreach ($entityCategoryAttributes as $entityCategoryAttribute) {
-			if (substr($entityCategoryAttribute, 0, strlen(self::EDUID_IDP_GROUP)) === self::EDUID_IDP_GROUP) {
-				$this->entityCategory = substr($entityCategoryAttribute, strlen(self::EDUID_IDP_GROUP), strlen($entityCategoryAttribute) - strlen(self::EDUID_IDP_GROUP));
-			}
 		}
 
 		$isHostelVerified = false;
@@ -84,19 +84,21 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
 		}
 
 		try {
-			$rpcConnector = (new sspmod_perun_AdapterRpc())->getConnector();
-			$this->cesnetEligibleLastSeen = $rpcConnector->get('attributesManager', 'getAttribute', array(
-				'user' => $user->getId(),
-				'attributeName' => $this->cesnetEligibleLastSeenAttr,
-			));
-
-			if ((!empty($this->eduPersonScopedAffiliation) && !is_null($this->entityCategory) && $this->isCesnetEligible())
-				|| $isHostelVerified) {
-				$this->cesnetEligibleLastSeen['value'] = date("Y-m-d H:i:s");
-				$rpcConnector->post('attributesManager', 'setAttribute', array(
+			if (!in_array($this->spEntityId, $this->listOfPerunEntityIds)) {
+				$this->cesnetEligibleLastSeen = $this->rpcConnector->get('attributesManager', 'getAttribute', array(
 					'user' => $user->getId(),
-					'attribute' => $this->cesnetEligibleLastSeen,
+					'attributeName' => $this->cesnetEligibleLastSeenAttr,
 				));
+			}
+
+			if ((!empty($this->eduPersonScopedAffiliation) && $this->isCesnetEligible()) || $isHostelVerified) {
+				$this->cesnetEligibleLastSeen['value'] = date("Y-m-d H:i:s");
+				if (!in_array($this->spEntityId, $this->listOfPerunEntityIds)) {
+					$this->rpcConnector->post('attributesManager', 'setAttribute', array(
+						'user' => $user->getId(),
+						'attribute' => $this->cesnetEligibleLastSeen,
+					));
+				}
 			}
 		} catch (Exception $ex) {
 			SimpleSAML\Logger::warning("cesnet:IsCesnetEligible - " . $ex->getMessage());
@@ -108,41 +110,39 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
 	}
 
 	/**
-	 * Return true if combination of user attributes and IdP metadata attributes pass through the eduid filter, False if not
-	 * @return bool True if combination of attributes pass through the filter, else False
+	 * Returns true if one of user's affiliation is in allowed affiliations for this IdP , False if not
+	 * @return bool
 	 */
 	private function isCesnetEligible() {
-		if ($this->entityCategory === self::UNIVERSITY) {
-			foreach ($this->eduPersonScopedAffiliation as $affiliation) {
-				if (preg_match("/(^employee@.+\.cz$)|(^faculty@.+\.cz$)|(^member@.+\.cz$)|(^student@.+\.cz$)|(^staff@.+\.cz$)/", $affiliation, $matches)) {
-					return true;
-				}
-			}
-		} elseif ($this->entityCategory === self::AVCR) {
-			foreach ($this->eduPersonScopedAffiliation as $affiliation) {
-				if (preg_match("/(^member@.+\.cz$)|(^staff@.+\.cz$)/", $affiliation, $matches)) {
-					return true;
-				}
-			}
-		} elseif ($this->entityCategory === self::LIBRARY) {
-			foreach ($this->eduPersonScopedAffiliation as $affiliation) {
-				if (preg_match("/(^employee@.+\.cz$)|(^staff@.+\.cz$)/", $affiliation, $matches)) {
-					return true;
-				}
-			}
-		} elseif ($this->entityCategory === self::HOSPITAL) {
-			foreach ($this->eduPersonScopedAffiliation as $affiliation) {
-				if (preg_match("/(^employee@.+\.cz$)|(^staff@.+\.cz$)/", $affiliation, $matches)) {
-					return true;
-				}
-			}
-		} elseif ($this->entityCategory === self::OTHER) {
-			foreach ($this->eduPersonScopedAffiliation as $affiliation) {
-				if (preg_match("/(^employee@.+\.cz$)|(^member@.+\.cz$)|(^employee@bbmri-eric\.eu$)|(^member@bbmri-eric\.eu$)/", $affiliation, $matches)) {
-					return true;
-				}
+
+		$allowedAffiliations = $this->getAllowedAffiliations($this->idpEntityId);
+		foreach ($this->eduPersonScopedAffiliation as $userAffiliation) {
+			$userAffiliationWithoutScope = explode("@", $userAffiliation)[0];
+			if (!is_null($userAffiliationWithoutScope) && !empty($userAffiliationWithoutScope)
+				&& in_array($userAffiliationWithoutScope, $allowedAffiliations)) {
+				return true;
 			}
 		}
 		return false;
+	}
+
+	/**
+	 * Return list of allowed affiliations for IdP from CESNET LDAP
+	 * @param $idpEntityId entityId of IdP
+	 * @return array of allowed affiliations
+	 */
+	private function getAllowedAffiliations($idpEntityId) {
+		$allowedAffiliations = array();
+
+		try {
+			$affiliations = $this->cesnetLdapConnector->searchForEntity(self::ORGANIZATION_LDAP_BASE,'(entityIDofIdP=' . $idpEntityId . ')', array(
+				'cesnetcustomeraffiliation'))['cesnetcustomeraffiliation'];
+			foreach ($affiliations as $affiliation) {
+				array_push($allowedAffiliations, $affiliation);
+			}
+		} catch (Exception $ex) {
+			SimpleSAML\Logger::warning("cesnet:IsCesnetEligible - Unable to connect to LDAP!");
+		}
+		return $allowedAffiliations;
 	}
 }